Move `StreamRecord` class to public module, add docs
